requirements.txt
anthropic>=0.25.0
requests>=2.31.0
pyyaml>=6.0
mkdocs-material>=9.5.0
